police background common component
